refactor: move job state machine in lib
