+++ src/spaceone/inventory/manager/job_task_manager.py
@@ -1,4 +1,3 @@
-import abc
 import logging
 import json
 from jsonschema import validate
@@ -9,12 +8,9 @@ from spaceone.core.manager import BaseManager
 from spaceone.core.scheduler.task_schema import SPACEONE_TASK_SCHEMA
 from spaceone.inventory.manager.job_manager import JobManager
 from spaceone.inventory.model.job_task_model import JobTask
-from spaceone.inventory.error import *
-
+from spaceone.inventory.lib.job_task_state import JobTaskStateMachine
 
 _LOGGER = logging.getLogger(__name__)
-
-MAX_MESSAGE_LENGTH = 2000
 
 
 class JobTaskManager(BaseManager):
@@ -181,99 +177,3 @@ class JobTaskManager(BaseManager):
         }
         _LOGGER.debug(f'[_create_task] tasks: {stp}')
         return stp
-
-
-PENDING = 'PENDING'
-CANCELED = 'CANCELED'
-INPROGRESS = 'IN_PROGRESS'
-SUCCESS = 'SUCCESS'
-FAILURE = 'FAILURE'
-
-
-class JobTaskState(metaclass=abc.ABCMeta):
-    def __init__(self):
-        self.handle()
-
-    @abc.abstractmethod
-    def handle(self):
-        pass
-
-
-class PendingState(JobTaskState):
-    def handle(self):
-        pass
-
-    def __str__(self):
-        return PENDING
-
-
-class CanceledState(JobTaskState):
-    def handle(self):
-        pass
-
-    def __str__(self):
-        return CANCELED
-
-
-class InprogressState(JobTaskState):
-    def handle(self):
-        pass
-
-    def __str__(self):
-        return INPROGRESS
-
-
-class SuccessState(JobTaskState):
-    def handle(self):
-        pass
-
-    def __str__(self):
-        return SUCCESS
-
-
-class FailureState(JobTaskState):
-    def handle(self):
-        pass
-
-    def __str__(self):
-        return FAILURE
-
-
-STATE_DIC = {
-    'PENDING': PendingState(),
-    'CANCELED': CanceledState(),
-    'IN_PROGRESS': InprogressState(),
-    'SUCCESS': SuccessState(),
-    'FAILURE': FailureState()
-}
-
-
-class JobTaskStateMachine(object):
-    def __init__(self, job_task_vo):
-        self.job_task_id = job_task_vo.job_task_id
-        self._state = STATE_DIC[job_task_vo.status]
-
-    def inprogress(self):
-        if isinstance(self._state, (PendingState, InprogressState, SuccessState, FailureState)):
-            self._state = InprogressState()
-        else:
-            raise ERROR_JOB_TASK_STATE_CHANGE(action='INPROGRESS', job_task_id=self.job_task_id, status=str(self._state))
-        return self.get_state()
-
-    def success(self):
-        if isinstance(self._state, (InprogressState)):
-            self._state = SuccessState()
-        else:
-            raise ERROR_JOB_TASK_STATE_CHANGE(action='SUCCESS', job_task_id=self.job_task_id, status=str(self._state))
-        return self.get_state()
-
-    def failure(self):
-        self._state = FailureState()
-        return self.get_state()
-
-    def canceled(self):
-        self._state = CanceledState()
-        return self.get_state()
-
-    def get_state(self):
-        return str(self._state)
